+++ .vscode/settings.json
@@ -13,7 +13,9 @@
     "helptask",
     "pingderby",
     "secondtask",
+    "setupderby",
     "targetchannel",
+    "taskname",
     "tasktitle",
     "togglemember"
   ]
